requirements.txt
openai
groq
mistralai
python-docx
markdown2